refactor(plugins) keymaps
